Add files via upload
This chart illustrates the makeup of the Lehigh undergraduate colleges over the past five years. We can see that the business school has fluctuated somewhat, with a slight increase then decrease in the beginning, but has stayed the same since. Engineering has been on a decreasing trend and arts and sciences has been increasing over the past five years.